Lecture 208 - Installing dependency libraries for our userspace and running the calculator

+++ PeachOS64Bit/programs/stdlib/src/file.c
@@ -1,16 +1,26 @@
+/*
+ * Copyright (C) 2025 Daniel McCarthy <daniel@dragonzap.com>
+ * Part of the PeachOS Part Two Development Series.
+ * https://github.com/nibblebits/PeachOS64BitCourse
+ * https://github.com/nibblebits/PeachOS64BitModuleTwo
+ * Licensed under the GNU General Public License version 2 (GPLv2).
+ *
+ * Community contributors to this source file:
+ * NONE AS OF YET
+ * ----------------
+ * Disclaimer: Contributors are hobbyists that contributed to the public source code, they are not affiliated or endorsed by Daniel McCarthy the author of the PeachOS Kernel      
+ * development video series. Contributors did not contribute to the video content or the teaching and have no intellectual property rights over the video content for the course video * material and did not contribute to the video material in anyway.
+ */
+
 #include "file.h"
 #include "peachos.h"
-#include <stdint.h>
-#include <stddef.h>
-
 int fopen(const char* filename, const char* mode)
 {
     return (int) peachos_fopen(filename, mode);
 }
-
 void fclose(int fd)
 {
-    peachos_fclose((size_t) fd);
+    peachos_fclose((long) fd);
 }
 
 int fread(void* buffer, size_t size, size_t count, long fd)
