+++ PeachOS64Bit/programs/stdlib/src/string.h
@@ -1,12 +1,28 @@
+/*
+ * Copyright (C) 2025 Daniel McCarthy <daniel@dragonzap.com>
+ * Part of the PeachOS Part Two Development Series.
+ * https://github.com/nibblebits/PeachOS64BitCourse
+ * https://github.com/nibblebits/PeachOS64BitModuleTwo
+ * Licensed under the GNU General Public License version 2 (GPLv2).
+ *
+ * Community contributors to this source file:
+ * NONE AS OF YET
+ * ----------------
+ * Disclaimer: Contributors are hobbyists that contributed to the public source code, they are not affiliated or endorsed by Daniel McCarthy the author of the PeachOS Kernel      
+ * development video series. Contributors did not contribute to the video content or the teaching and have no intellectual property rights over the video content for the course video * material and did not contribute to the video material in anyway.
+ */
+
 #ifndef PEACHOS_STRING_H
 #define PEACHOS_STRING_H
 
 #include <stdbool.h>
+#include <stdint.h>
+#include <stddef.h>
 
 char tolower(char s1);
-int strlen(const char* ptr);
-int strnlen(const char* ptr, int max);
-int strnlen_terminator(const char* str, int max, char terminator);
+size_t strlen(const char* ptr);
+size_t strnlen(const char* ptr, size_t max);
+size_t strnlen_terminator(const char* str, size_t max, char terminator);
 int istrncmp(const char* s1, const char* s2, int n);
 int strncmp(const char* str1, const char* str2, int n);
 char* strcpy(char* dest, const char* src);
